Rename chain_info mod to constants following op-revm
