+++ crates/celo-revm/src/lib.rs
@@ -6,7 +6,7 @@
 extern crate alloc as std;
 
 pub mod api;
-pub mod chain_info;
+pub mod constants;
 pub mod evm;
 pub mod handler;
 pub mod precompiles;
